refactor(MovieReviewDetailModal): 버튼 클릭에 대한 책임 분리
1. ModalCloseButton 컴포넌트 구현

2. MovieReviewDetailModal 내 버튼 클릭 핸들러 및 이벤트 리스너를 ModalCloseButton에 위임
